ai-service/pyproject.toml
[project]
name = "docs-to-testcases-ai"
version = "1.0.0"
description = "AI service for document processing, RAG, and test case generation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "chromadb>=0.5.0",
    "httpx>=0.27.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.11.0",
    "markdown>=3.6",
    "beautifulsoup4>=4.12.0",
    "langchain-text-splitters>=0.2.0",
]
